Fix get_fields lookup
Some other plugins that work with django rest framework need the get_fields method to do API introspection (django-rest-swagger)  This allows for that to continue to work without breaking because marshmallow has different ideas about serailizers read_only/write_only then django rest frameworks model serializers as it pertains to fields.  Also marshmallow's default class object of "Missing" does not serialize to JSON properly for django-rest-swagger.  This is the fix for that, but maybe it should actually be propegated down into that class in marshmallow.